# dist/api.yaml
openapi: 3.0.3
info:
  version: 1.0.0
  title: Boilerplate
  description: Multi-file boilerplate for OpenAPI Specification.
servers:
  - url: 'http://localhost:8080/v1'
    description: Локальный сервер
paths:
  /auth/login:
    post:
      security:
        - origin:
            - '*'
      description: Login
      operationId: Login
      tags:
        - Auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/login'
      responses:
        '200':
          description: Logged
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/tokenReponse'
        '400':
          description: Bad request
  /university/add-students:
    post:
      description: Add new students
      operationId: AddStudents
      tags:
        - University
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/addStudent'
      responses:
        '200':
          description: Students added
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/creationResult'
        '400':
          description: Bad request
        '401':
          description: Not Unauthorized
components:
  securitySchemes:
    authorization:
      type: apiKey
      in: header
      name: Authorization
    origin:
      type: apiKey
      in: header
      name: Origin
  parameters:
    id:
      name: id
      in: path
      description: ID
      required: true
      schema:
        type: integer
        format: int
  schemas:
    login:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          description: User email
        password:
          type: string
          description: User password
    tokenReponse:
      type: object
      required:
        - accessToken
        - expiresIn
      properties:
        accessToken:
          type: string
        expiresIn:
          type: integer
          x-go-type: int64
    student:
      type: object
      required:
        - email
        - firstName
        - lastName
        - middleName
        - phoneNumber
      properties:
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        middleName:
          type: string
        phoneNumber:
          type: string
    addStudent:
      type: object
      required:
        - universityId
        - students
      properties:
        universityId:
          type: integer
        students:
          type: array
          items:
            $ref: '#/components/schemas/student'
    creationResult:
      type: object
      required:
        - studentEmail
        - isCreated
      properties:
        studentEmail:
          type: string
        isCreated:
          type: boolean